[level 4] Title: 징검다리, Time: 44.30 ms, Memory: 79.8 MB -BaekjoonHub

+++ 프로그래머스/4/43236. 징검다리/징검다리.java
@@ -13,7 +13,7 @@ class Solution {
             int prev = 0;
             
             for (int rock : rocks) {
-                long dis = rock - prev;
+                int dis = rock - prev;
                 if (dis < mid) {
                     remove++;
                     if (remove > n) break;
